refactor(sync): rename sync_daemon.py -> hive_sync_daemon.py (item 9)
A generic "sync_daemon" process/module name risks colliding with unrelated
tooling. Rename to the hive-namespaced, descriptive hive_sync_daemon.py and keep
a thin sync_daemon.py compat shim (re-exports the daemon) for ONE release so a
service/cron entry still pointing at the old path keeps working until
`hive-mind update` rewrites the units.

Updated all launch/path references across the service backends (_units/_launchd/
_runit), installer, uninstaller, sync_smoke.sh, ci.yml, docs, and tests; tests now
load the real module (the shim's `import *` can't re-export underscore names). The
`hv sync daemon` CLI and its orphan-detection (matches the `hv sync daemon` argv)
are unchanged; the crontab `grep -v sync_daemon` stays broad to match both names
during the transition.
